+++ client/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added edit functionality for names ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
+- Added set default functionality for My Collections ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 
 ### Changed